Move rational_design package scripts and fix import names
